Replaced longs with ints for time intervals and timeouts

+++ io-mokamint-application-integration/src/test/java/io/mokamint/application/integration/tests/RemoteApplicationTests.java
@@ -86,7 +86,7 @@ public class RemoteApplicationTests extends AbstractLoggedTests {
 		}
 	}
 
-	private final static long TIME_OUT = 2000L;
+	private final static int TIME_OUT = 2000;
 
 	/**
 	 * Test server implementation.
